refactor(OMN-169): split script-response-schemas.ts into common/read/analyze/write modules
Behavior-preserving module split of the 1635-line script-response-schemas.ts.
The file becomes a thin re-export barrel; all 23 import sites are untouched and
the public export surface is identical (43 names, verified by diff).

- response-schemas/common.ts: shared leaves (isoDate/isoDateOrNull/warningsArray)
  + envelope factories (v3EnvelopeSchema/astEnvelopeSchema/listResultSchema/
  reviewSuccessSchema/V3EnvelopeSuccessSchema)
- response-schemas/{read,analyze,write}.ts: domain schemas, importing only ./common
- no cross-domain references, no circular imports

Correctness oracle: full unit suite (3417 tests) + tsc build, both green.

Closes OMN-169
